Updated contradictory test cases 70, 73, and 76
There were contradictions in the test cases in test/properties.js
Test case 7 suggests that duplicate keys should be treated by overwriting keys defined toward the top of the document with keys found lower in the document.

Test cases 70, 73, and 76 suggest that duplicate keys should signify an array of values.

Test case 7 seems to be the generally accepted standard for properties files

After this commit, an array of values will be treated by using their indices as key values.

For example:

{ "test": [ "this", "array" ] }

will be serialized into:

test.0=this
test.1=array

This can be changed to

test[0]=this
test[1]=array

if need be, but it seems wrong to serialize it to

test=this
test=array

as test cases 70, 73, and 76 imply.

Update:

updated test cases 70, 73, and 76 so that they expect

{ "test": [ "this", "array" ] }

to be serialized to

test[0]=this
test[1]=array